requirements.txt
# requirements.txt

Flask
Flask-SQLAlchemy
PyJWT
mysql-connector-python
python-dotenv
requests